Move locked figures files to their own directories. Update references in consuming files.
