+++ packages/perseus-editor/src/components/__tests__/locked-figures-section.test.tsx
@@ -3,7 +3,7 @@ import {screen, render} from "@testing-library/react";
 import {userEvent as userEventLib} from "@testing-library/user-event";
 import * as React from "react";
 
-import LockedFiguresSection from "../locked-figures-section";
+import LockedFiguresSection from "../graph-locked-figures/locked-figures-section";
 import {getDefaultFigureForType} from "../util";
 
 import type {UserEvent} from "@testing-library/user-event";
