+++ packages/perseus-editor/src/components/__tests__/locked-line-settings.test.tsx
@@ -3,7 +3,7 @@ import {render, screen} from "@testing-library/react";
 import {userEvent as userEventLib} from "@testing-library/user-event";
 import * as React from "react";
 
-import LockedLineSettings from "../locked-line-settings";
+import LockedLineSettings from "../graph-locked-figures/locked-line-settings";
 import {getDefaultFigureForType} from "../util";
 
 import type {UserEvent} from "@testing-library/user-event";
