+++ packages/perseus-editor/src/components/graph-locked-figures/locked-ellipse-settings.tsx
@@ -7,10 +7,10 @@ import {LabelMedium, LabelLarge} from "@khanacademy/wonder-blocks-typography";
 import {StyleSheet} from "aphrodite";
 import * as React from "react";
 
-import AngleInput from "./angle-input";
-import ColorSelect from "./color-select";
-import CoordinatePairInput from "./coordinate-pair-input";
-import EllipseSwatch from "./ellipse-swatch";
+import AngleInput from "../angle-input";
+import ColorSelect from "../color-select";
+import CoordinatePairInput from "../coordinate-pair-input";
+import EllipseSwatch from "../ellipse-swatch";
 import LockedFigureSettingsAccordion from "./locked-figure-settings-accordion";
 import LockedFigureSettingsActions from "./locked-figure-settings-actions";
