+++ packages/perseus-editor/src/components/graph-locked-figures/locked-figure-select.tsx
