+++ packages/perseus-editor/src/components/graph-locked-figures/locked-figures-section.tsx
@@ -12,10 +12,10 @@ import * as React from "react";
 
 import LockedFigureSelect from "./locked-figure-select";
 import LockedFigureSettings from "./locked-figure-settings";
-import {getDefaultFigureForType} from "./util";
+import {getDefaultFigureForType} from "../util";
 
 import type {LockedFigureSettingsMovementType} from "./locked-figure-settings-actions";
-import type {Props as InteractiveGraphEditorProps} from "../widgets/interactive-graph-editor";
+import type {Props as InteractiveGraphEditorProps} from "../../widgets/interactive-graph-editor";
 import type {LockedFigure, LockedFigureType} from "@khanacademy/perseus";
 
 type Props = {
