+++ packages/perseus/src/widgets/interactive-graphs/locked-figures/locked-polygon.tsx
@@ -1,9 +1,12 @@
 import {Point, Polygon} from "mafs";
 import * as React from "react";
 
-import {lockedFigureColors, lockedFigureFillStyles} from "../../perseus-types";
+import {
+    lockedFigureColors,
+    lockedFigureFillStyles,
+} from "../../../perseus-types";
 
-import type {LockedPolygonType} from "../../perseus-types";
+import type {LockedPolygonType} from "../../../perseus-types";
 
 const LockedPolygon = (props: LockedPolygonType) => {
     const {points, color, showVertices, fillStyle, strokeStyle} = props;
