+++ packages/perseus/src/widgets/interactive-graphs/locked-figures/locked-vector.tsx
@@ -1,10 +1,9 @@
 import * as React from "react";
 
-import {lockedFigureColors} from "../../perseus-types";
+import {lockedFigureColors} from "../../../perseus-types";
+import {Vector} from "../graphs/components/vector";
 
-import {Vector} from "./graphs/components/vector";
-
-import type {LockedVectorType} from "../../perseus-types";
+import type {LockedVectorType} from "../../../perseus-types";
 
 const LockedVector = (props: LockedVectorType) => {
     const {color, points} = props;
